Added trait to unit tests, now passing so removed bug warning from readme
