+++ tests/Bravo3/Orm/Tests/Entities/Product.php
@@ -12,18 +12,14 @@ use Doctrine\Common\Annotations\Annotation\IgnoreAnnotation;
  */
 class Product
 {
+    use NameTrait;
+
     /**
      * @var int
      * @Orm\Id
      * @Orm\Column(type="int")
      */
     protected $id;
-
-    /**
-     * @var string
-     * @Orm\Column(type="string")
-     */
-    protected $name;
 
     /**
      * @var string
@@ -85,28 +81,6 @@ class Product
     public function setId($id)
     {
         $this->id = $id;
-        return $this;
-    }
-
-    /**
-     * Get Name
-     *
-     * @return string
-     */
-    public function getName()
-    {
-        return $this->name;
-    }
-
-    /**
-     * Set Name
-     *
-     * @param string $name
-     * @return $this
-     */
-    public function setName($name)
-    {
-        $this->name = $name;
         return $this;
     }
